+++ package.json
@@ -15,7 +15,7 @@
     "type": "git",
     "url": "https://github.com/7finney/sol-exec"
   },
-  "version": "0.0.2",
+  "version": "0.0.3",
   "engines": {
     "vscode": "^1.77.0"
   },
